Refactor bot into smaller modules
