+++ cogs/helpformatter.py
@@ -4,7 +4,7 @@ from discord.ext import commands
 import re
 
 # Bot Utilities
-from cogs.utils.paginator import HelpPaginator, Scroller
+from .utils.userinteraction import HelpPaginator, Scroller
 
 usermention = r"<@!?\d{17,19}>"
