+++ cogs/music/basic_commands.py
@@ -10,8 +10,7 @@ import asyncio
 import re
 
 from ..utils import checks, timeformatter
-from ..utils.paginator import QueuePaginator, Scroller
-from ..utils.selector import Selector
+from ..utils.userinteraction import QueuePaginator, Scroller, Selector
 from .decorators import require_playing, require_queue, require_voice_connection, voteable
 
 time_rx = re.compile('[0-9]+')
