+++ cogs/music/extra_commands.py
@@ -15,7 +15,7 @@ import urllib
 from bs4 import BeautifulSoup
 
 from ..utils import checks
-from ..utils.paginator import Scroller, TextPaginator
+from ..utils.userinteraction import Scroller, TextPaginator
 from .decorators import require_playing, require_voice_connection, voteable
