+++ cogs/utils/localisation/localizedcontext.py
@@ -2,7 +2,7 @@
 from discord.ext import commands
 
 
-class Context(commands.Context):
+class LocalizedContext(commands.Context):
     def __init__(self, **kwargs):
         super().__init__(**kwargs)
         if self.message.guild:
