+++ cogs/utils/localisation/localizer/__init__.py
@@ -1,0 +1,4 @@
+from .localizer import Localizer
+from .localizerwrapper import LocalizerWrapper
+
+__all__ = ['LocalizerWrapper', 'Localizer']
