+++ cogs/utils/settingsmanager/__init__.py
@@ -1,0 +1,3 @@
+from .settingsmanager import Settings
+
+__all__ = ['Settings']
