+++ cogs/utils/userinteraction/__init__.py
@@ -1,0 +1,6 @@
+from .scroller import Scroller
+from .selector import Selector
+from .paginators import BasePaginator, TextPaginator, FieldPaginator, HelpPaginator, QueuePaginator, CantScrollException
+
+__all__ = ['Scroller', 'Selector', 'QueuePaginator', 'HelpPaginator',
+           'TextPaginator', 'FieldPaginator', 'CantScrollException', 'BasePaginator']
