+++ cogs/utils/userinteraction/paginators/__init__.py
@@ -1,0 +1,7 @@
+from .queuepaginator import QueuePaginator
+from .helppaginator import HelpPaginator
+from .textpaginator import TextPaginator
+from .fieldpaginator import FieldPaginator
+from .basepaginator import CantScrollException, BasePaginator
+
+__all__ = ['QueuePaginator', 'HelpPaginator', 'TextPaginator', 'FieldPaginator', 'CantScrollException', 'BasePaginator']
