+++ cogs/utils/userinteraction/selector.py
@@ -4,7 +4,7 @@ import discord
 import asyncio
 import inspect
 
-from .paginator import TextPaginator
+from .paginators import TextPaginator
 
 
 class Selector(TextPaginator):
